puzzle-api-python/requirements.txt
flask
flask-restful
flask-cors
pyjwt
mysql-connector-python
jsonpickle
bcrypt
